測試調整

+++ NT_AirPollution.Web/Controllers/FormController.cs
@@ -39,7 +39,7 @@ namespace NT_AirPollution.Web.Controllers
 
         public JsonResult GetTotalMoney(FormView form)
         {
-            int S_AMT = _formService.CalcTotalMoney(form, 0);
+            double S_AMT = _formService.CalcTotalMoney(form, 0);
             return Json(new AjaxResult { Status = true, Message = S_AMT }, JsonRequestBehavior.AllowGet);
         }
